requirements.txt
tensorflow
opencv-python-headless
pillow
matplotlib
numpy
streamlit

